GH #25039 Facelets in Console: Clean up
Moved CDI producers into Admin Core, so that they can be reused by other modules and we avoid conflicts or implicit dependencies if multiple modules define the same producers or rely on beans produced by other modules.
