+++ appserver/admingui/commandrecorder/src/main/resources/META-INF/resources/commandrecorder/plugin/mastheadStatusArea.jsf
@@ -1,5 +1,4 @@
 <!--
-
     Copyright (c) 2024 Contributors to the Eclipse Foundation
 
     This program and the accompanying materials are made available under the
@@ -13,7 +12,6 @@
     https://www.gnu.org/software/classpath/license.html.
 
     SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-
 -->
 
 <f:verbatim>
