+++ appserver/admingui/war/src/main/java/org/glassfish/admingui/common/AdminGuiViewHandler.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
- * Copyright (c) 2010, 2018 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
