+++ appserver/admingui/war/src/main/webapp/WEB-INF/web.xml
@@ -113,6 +113,10 @@
     </servlet-mapping>
     <servlet-mapping>
         <servlet-name>FacesServlet</servlet-name>
+        <url-pattern>/faces/*</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>FacesServlet</servlet-name>
         <url-pattern>*.xhtml</url-pattern>
     </servlet-mapping>
     <servlet-mapping>
