--- appserver/admingui/war/src/main/webapp/iframe.xhtml
@@ -1,18 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:h="jakarta.faces.html">
-    <h:head>
-        <title>Facelet Title</title>
-        <link rel="stylesheet" type="text/css" href="#{request.contextPath}/theme/com/sun/webui/jsf/suntheme/css/css_master.css"/>
-        <link rel="stylesheet" type="text/css" href="#{request.contextPath}/resources/yui/reset-fonts-grids.css"/>
-        <link rel="stylesheet" type="text/css" href="#{request.contextPath}/resources/yui/assets/skins/sam/resize.css"/>
-        <link rel="stylesheet" type="text/css" href="#{request.contextPath}/resources/yui/assets/skins/sam/layout.css"/>
-        <link rel="stylesheet" type="text/css" href="#{request.contextPath}/resource/common/css/style.css"/>
-    </h:head>
-    <h:body>
-        <div style="text-align: left; height: 20px; background-color: #5b87a4">
-            <div class="mastheadButton_4_sun4" style="margin: 5px;"><a href="#">X #{headerView.title}</a></div>
-        </div>
-    </h:body>
-</html>
